+++ src/components/KeyAgencyNavigation.vue
@@ -2,13 +2,10 @@
   <v-row>
     <v-col md="6">
       <v-btn text color="primary" @click='$emit("fromKeyAgencyToNationalFramework")'>
-        <v-icon>mdi-step-backward</v-icon>
-        National framework
-      </v-btn>
+        <v-icon>mdi-step-backward</v-icon>previous</v-btn>
     </v-col>
     <v-col md="6" class="text-right">
-      <v-btn text color="primary" @click='$emit("fromKeyAgencyToEvolution")'>Evolution
-        <v-icon>mdi-step-forward</v-icon>
+      <v-btn text color="primary" @click='$emit("fromKeyAgencyToEvolution")'>next<v-icon>mdi-step-forward</v-icon>
       </v-btn>
     </v-col>
   </v-row>